Fixed XMouse Control 'sticky key' bug
When holding Mod Key, if Mod is released before Key, then the Mod Key
key release event won't be triggered and XMouseControl won't receive the
signal to stop controlling the mouse. Now we remove all control when the
modifier is pressed.

+++ src/Extensions/xmousecontrol.h
@@ -75,13 +75,13 @@ void addXMouseControlMask(int mask);
  *
  * @param multiplier
  */
-void adjustSpeed(double multiplier);
+void adjustSpeed(int multiplier);
 /**
  * Increases how many times the scroll button is pressed every update
  *
  * @param multiplier
  */
-void adjustScrollSpeed(double multiplier);
+void adjustScrollSpeed(int multiplier);
 /**
  * Applies a single update
  */
